Fixed problem with page_scale for PDF files
git-svn-id: svn://svn.code.sf.net/p/osra/code/trunk@736 049e15af-3938-0410-b28b-82a5db9f179f

+++ src/osra.cpp
@@ -5442,7 +5442,7 @@ int main(int argc, char **argv)
       density << input_resolution << "x" << input_resolution;
       image.density(density.str());
 
-      if (type == "PDF" || type == "PS") page_scale*=72/input_resolution;
+      if (type == "PDF" || type == "PS") page_scale*=(double) 72/input_resolution;
 
 
 #ifdef ANDROID
